Use the Engine Pi instead of the Engine Alpha

+++ src/main/java/rocks/friedrich/jwinf/blockly_robot/gui/Painter.java
@@ -1,12 +1,12 @@
 package rocks.friedrich.jwinf.blockly_robot.gui;
 
-import ea.Scene;
-import ea.actor.Text;
+import de.pirckheimer_gymnasium.engine_pi.Scene;
+import de.pirckheimer_gymnasium.engine_pi.actor.Text;
 import rocks.friedrich.jwinf.blockly_robot.logic.level.Difficulty;
 
 public class Painter
 {
-    public static void paintVersionHeading(Scene scene, float x, float y,
+    public static void paintVersionHeading(Scene scene, double x, double y,
             Difficulty difficulty)
     {
         Text text = TextMaker.createText(
