+++ src/main/java/rocks/friedrich/jwinf/blockly_robot/gui/map/CoordinateSystemTranslator.java
@@ -1,6 +1,6 @@
 package rocks.friedrich.jwinf.blockly_robot.gui.map;
 
-import ea.Vector;
+import de.pirckheimer_gymnasium.engine_pi.Vector;
 import rocks.friedrich.jwinf.blockly_robot.logic.navigation.Coords;
 
 public class CoordinateSystemTranslator
@@ -19,15 +19,15 @@ public class CoordinateSystemTranslator
      * Die x-Koordinate des linken unteren Ecks, an dem das Gitter im
      * Engine-Alpha-Koordinatensystem verankert ist.
      */
-    public float x;
+    public double x;
 
     /**
      * Die y-Koordinate des linken unteren Ecks, an dem das Gitter im
      * Engine-Alpha-Koordinatensystem verankert ist.
      */
-    public float y;
+    public double y;
 
-    public CoordinateSystemTranslator(int rows, int cols, float x, float y)
+    public CoordinateSystemTranslator(int rows, int cols, double x, double y)
     {
         this.rows = rows;
         this.cols = cols;
@@ -41,7 +41,7 @@ public class CoordinateSystemTranslator
         this.cols = cols;
     }
 
-    public void setPosition(float x, float y)
+    public void setPosition(double x, double y)
     {
         this.x = x;
         this.y = y;
@@ -50,27 +50,27 @@ public class CoordinateSystemTranslator
     /**
      * y-Koordinate des Ursprungs des Gitters (links oben)
      */
-    public float row0y()
+    public double row0y()
     {
         return y + rows - 1;
     }
 
-    public int toRow(float y)
+    public int toRow(double y)
     {
-        return Math.round(row0y() - y);
+        return (int) Math.round(row0y() - y);
     }
 
-    public int toCol(float x)
+    public int toCol(double x)
     {
-        return Math.round(x - this.x);
+        return (int) Math.round(x - this.x);
     }
 
-    public float toX(int col)
+    public double toX(int col)
     {
         return col + x;
     }
 
-    public float toY(int row)
+    public double toY(int row)
     {
         return row0y() - row;
     }
@@ -83,7 +83,7 @@ public class CoordinateSystemTranslator
         return new Coords(toRow(vector.getY()), toCol(vector.getX()));
     }
 
-    public Coords toPoint(float x, float y)
+    public Coords toPoint(double x, double y)
     {
         return toPoint(new Vector(x, y));
     }
